allow tests to run in two modes

+++ test/message_extraction/message_extraction_test.dart
@@ -103,22 +103,30 @@ void copyFilesToTempDirectory() {
     asTestDirPath('mock_flutter/services.dart'),
   ];
 
-  for (var filename in files) {
-    var file = new File(filename);
+  for (var filePath in files) {
+    var file = new File(filePath);
     if (file.existsSync()) {
-      file.copySync(path.join(tempDir, path.basename(filename)));
+      file.copySync(path.join(tempDir, path.basename(filePath)));
     }
   }
 
-  // Copy our package_config.json file so the test can locate packages.
-  var sourcePackageConfig =
-      File(path.join('.dart_tool', 'package_config.json'));
-  var destPackageConfig =
-      File(path.join(tempDir, '.dart_tool', 'package_config.json'));
-  if (!destPackageConfig.parent.existsSync()) {
-    destPackageConfig.parent.createSync();
+  // TODO:devoncarew): Improve the integration testing story here.
+  // For github, copy the package_config.json file so the test can locate
+  // packages.
+  if (Directory('.dart_tool').existsSync()) {
+    var sourcePackageConfig =
+        File(path.join('.dart_tool', 'package_config.json'));
+    var destPackageConfig =
+        File(path.join(tempDir, '.dart_tool', 'package_config.json'));
+    if (!destPackageConfig.parent.existsSync()) {
+      destPackageConfig.parent.createSync();
+    }
+    sourcePackageConfig.copySync(destPackageConfig.path);
+  } else {
+    // For google3, copy the .packages file.
+    var file = File(asTestDirPath('.packages'));
+    file.copySync(path.join(tempDir, path.basename(file.path)));
   }
-  sourcePackageConfig.copySync(destPackageConfig.path);
 }
 
 void deleteGeneratedFiles() {
